size_t

+++ src/GOF.cpp
@@ -1,7 +1,7 @@
 /***
  * @Author: ChenRP07
  * @Date: 2022-05-13 10:58:55
- * @LastEditTime: 2022-05-13 10:58:55
+ * @LastEditTime: 2022-05-27 20:35:29
  * @LastEditors: ChenRP07
  * @Description:
  */
@@ -15,23 +15,23 @@ using namespace pco::octree;
  * @param {int} group_of_frames
  * @return {*}
  */
-GOF::GOF(const int group_of_frames) : kGroupOfFrames{ group_of_frames } {
-    try {
-        if (this->kGroupOfFrames < 1) {
-            throw "GOF size is smaller than 1.";
-        }
-    }
-    catch (const char* error_message) {
-        std::cerr << "GOF constructing failed : " << error_message << std::endl;
-        std::exit(1);
-    }
+GOF::GOF(const size_t group_of_frames) : kGroupOfFrames{group_of_frames} {
+	try {
+		if (this->kGroupOfFrames < 1) {
+			throw "GOF size is smaller than 1.";
+		}
+	}
+	catch (const char* error_message) {
+		std::cerr << "GOF constructing failed : " << error_message << std::endl;
+		std::exit(1);
+	}
 
-    this->min_x_ = FLT_MAX;
-    this->max_x_ = FLT_MIN;
-    this->min_y_ = FLT_MAX;
-    this->max_y_ = FLT_MIN;
-    this->min_z_ = FLT_MAX;
-    this->max_z_ = FLT_MIN;
+	this->min_x_ = FLT_MAX;
+	this->max_x_ = FLT_MIN;
+	this->min_y_ = FLT_MAX;
+	this->max_y_ = FLT_MIN;
+	this->min_z_ = FLT_MAX;
+	this->max_z_ = FLT_MIN;
 }
 
 /***
@@ -40,18 +40,18 @@ GOF::GOF(const int group_of_frames) : kGroupOfFrames{ group_of_frames } {
  * @return {size_t} frame_patches_.size()
  */
 size_t GOF::size() const {
-    try {
-        if (this->frame_patches_.size() == this->motion_vectors_.size()) {
-            return this->frame_patches_.size();
-        }
-        else {
-            throw "Unmatching size between frames and matrices.";
-        }
-    }
-    catch (const char* error_message) {
-        std::cerr << "GOF is broken : " << error_message << std::endl;
-        std::exit(1);
-    }
+	try {
+		if (this->frame_patches_.size() == this->motion_vectors_.size()) {
+			return this->frame_patches_.size();
+		}
+		else {
+			throw "Unmatching size between frames and matrices.";
+		}
+	}
+	catch (const char* error_message) {
+		std::cerr << "GOF is broken : " << error_message << std::endl;
+		std::exit(1);
+	}
 }
 
 /***
@@ -60,18 +60,18 @@ size_t GOF::size() const {
  * @return {size_t} frame_patches_[index].size()
  */
 size_t GOF::size(const size_t __index) const {
-    try {
-        if (__index < this->frame_patches_.size()) {
-            return this->frame_patches_[__index].size();
-        }
-        else {
-            throw "Out of range.";
-        }
-    }
-    catch (const char* error_message) {
-        std::cerr << "Fatal error from function GOF::size(__index) : " << error_message << std::endl;
-        std::exit(1);
-    }
+	try {
+		if (__index < this->frame_patches_.size()) {
+			return this->frame_patches_[__index].size();
+		}
+		else {
+			throw "Out of range.";
+		}
+	}
+	catch (const char* error_message) {
+		std::cerr << "Fatal error from function GOF::size(__index) : " << error_message << std::endl;
+		std::exit(1);
+	}
 }
 
 /***
@@ -81,23 +81,23 @@ size_t GOF::size(const size_t __index) const {
  * @return {PointXYZRGB&} frame_patches_[__x][__y]
  */
 pcl::PointXYZRGB& GOF::operator()(const size_t __x, const size_t __y) {
-    try {
-        if (__x < this->frame_patches_.size()) {
-            if (__y < this->frame_patches_[__x].size()) {
-                return this->frame_patches_[__x][__y];
-            }
-            else {
-                throw "Second index out of range.";
-            }
-        }
-        else {
-            throw "First index out of range.";
-        }
-    }
-    catch (const char* error_message) {
-        std::cerr << "Fatal error from access GOF point : " << error_message << std::endl;
-        std::exit(1);
-    }
+	try {
+		if (__x < this->frame_patches_.size()) {
+			if (__y < this->frame_patches_[__x].size()) {
+				return this->frame_patches_[__x][__y];
+			}
+			else {
+				throw "Second index out of range.";
+			}
+		}
+		else {
+			throw "First index out of range.";
+		}
+	}
+	catch (const char* error_message) {
+		std::cerr << "Fatal error from access GOF point : " << error_message << std::endl;
+		std::exit(1);
+	}
 }
 
 /***
@@ -107,23 +107,23 @@ pcl::PointXYZRGB& GOF::operator()(const size_t __x, const size_t __y) {
  * @return {PointXYZRGB&} frame_patches_[__x][__y]
  */
 const pcl::PointXYZRGB& GOF::operator()(const size_t __x, const size_t __y) const {
-    try {
-        if (__x < this->frame_patches_.size()) {
-            if (__y < this->frame_patches_[__x].size()) {
-                return this->frame_patches_[__x][__y];
-            }
-            else {
-                throw "Second index out of range.";
-            }
-        }
-        else {
-            throw "First index out of range.";
-        }
-    }
-    catch (const char* error_message) {
-        std::cerr << "Fatal error from access GOF point : " << error_message << std::endl;
-        std::exit(1);
-    }
+	try {
+		if (__x < this->frame_patches_.size()) {
+			if (__y < this->frame_patches_[__x].size()) {
+				return this->frame_patches_[__x][__y];
+			}
+			else {
+				throw "Second index out of range.";
+			}
+		}
+		else {
+			throw "First index out of range.";
+		}
+	}
+	catch (const char* error_message) {
+		std::cerr << "Fatal error from access GOF point : " << error_message << std::endl;
+		std::exit(1);
+	}
 }
 
 /***
@@ -133,37 +133,37 @@ const pcl::PointXYZRGB& GOF::operator()(const size_t __x, const size_t __y) cons
  * @return {*}
  */
 void GOF::AddPatch(const pcl::PointCloud<pcl::PointXYZRGB>& __patch, const Eigen::Matrix4f& __matrix) {
-    try {
-        // frame_patches_'s size must be less than kGroupOfFrames
-        if (this->frame_patches_.size() < this->kGroupOfFrames) {
-            // add a patch
-            this->frame_patches_.emplace_back(pcl::PointCloud<pcl::PointXYZRGB>());
-            this->frame_patches_[this->frame_patches_.size() - 1].resize(__patch.size());
-            for (size_t i = 0; i < __patch.size(); i++) {
-                this->frame_patches_[this->frame_patches_.size() - 1][i] = __patch[i];
-                // update range of coordinates
-                this->min_x_ = this->min_x_ < __patch[i].x ? this->min_x_ : __patch[i].x;
-                this->max_x_ = this->max_x_ > __patch[i].x ? this->max_x_ : __patch[i].x;
-                this->min_y_ = this->min_y_ < __patch[i].y ? this->min_y_ : __patch[i].y;
-                this->max_y_ = this->max_y_ > __patch[i].y ? this->max_y_ : __patch[i].y;
-                this->min_z_ = this->min_z_ < __patch[i].z ? this->min_z_ : __patch[i].z;
-                this->max_z_ = this->max_z_ > __patch[i].z ? this->max_z_ : __patch[i].z;
-            }
-            // add transformation matrix
-            if (this->frame_patches_.size() == 1) {
-                this->motion_vectors_.emplace_back(Eigen::Matrix4f::Identity());
-            }
-            else {
-                this->motion_vectors_.emplace_back(__matrix);
-            }
-        }
-        else {
-            throw "Out of GOF size limitation.";
-        }
-    }
-    catch (const char* error_message) {
-        std::cerr << "Adding patch into GOF failed : " << error_message << std::endl;
-    }
+	try {
+		// frame_patches_'s size must be less than kGroupOfFrames
+		if (this->frame_patches_.size() < this->kGroupOfFrames) {
+			// add a patch
+			this->frame_patches_.emplace_back(pcl::PointCloud<pcl::PointXYZRGB>());
+			this->frame_patches_[this->frame_patches_.size() - 1].resize(__patch.size());
+			for (size_t i = 0; i < __patch.size(); i++) {
+				this->frame_patches_[this->frame_patches_.size() - 1][i] = __patch[i];
+				// update range of coordinates
+				this->min_x_ = this->min_x_ < __patch[i].x ? this->min_x_ : __patch[i].x;
+				this->max_x_ = this->max_x_ > __patch[i].x ? this->max_x_ : __patch[i].x;
+				this->min_y_ = this->min_y_ < __patch[i].y ? this->min_y_ : __patch[i].y;
+				this->max_y_ = this->max_y_ > __patch[i].y ? this->max_y_ : __patch[i].y;
+				this->min_z_ = this->min_z_ < __patch[i].z ? this->min_z_ : __patch[i].z;
+				this->max_z_ = this->max_z_ > __patch[i].z ? this->max_z_ : __patch[i].z;
+			}
+			// add transformation matrix
+			if (this->frame_patches_.size() == 1) {
+				this->motion_vectors_.emplace_back(Eigen::Matrix4f::Identity());
+			}
+			else {
+				this->motion_vectors_.emplace_back(__matrix);
+			}
+		}
+		else {
+			throw "Out of GOF size limitation.";
+		}
+	}
+	catch (const char* error_message) {
+		std::cerr << "Adding patch into GOF failed : " << error_message << std::endl;
+	}
 }
 
 /***
@@ -174,22 +174,22 @@ void GOF::AddPatch(const pcl::PointCloud<pcl::PointXYZRGB>& __patch, const Eigen
  * @return {*}
  */
 void GOF::GetPatch(pcl::PointCloud<pcl::PointXYZRGB>& __patch, Eigen::Matrix4f& __matrix, const size_t __index) const {
-    try {
-        if (__index < this->frame_patches_.size()) {
-            __patch.resize(this->frame_patches_[__index].size());
-            for (size_t i = 0; i < this->frame_patches_[__index].size(); i++) {
-                __patch[i] = this->frame_patches_[__index][i];
-            }
-            __matrix = this->motion_vectors_[__index];
-        }
-        else {
-            throw "Out of patch range.";
-        }
-    }
-    catch (const char* error_message) {
-        std::cerr << "Fatal error from function GetPatch() : " << error_message << std::endl;
-        std::exit(1);
-    }
+	try {
+		if (__index < this->frame_patches_.size()) {
+			__patch.resize(this->frame_patches_[__index].size());
+			for (size_t i = 0; i < this->frame_patches_[__index].size(); i++) {
+				__patch[i] = this->frame_patches_[__index][i];
+			}
+			__matrix = this->motion_vectors_[__index];
+		}
+		else {
+			throw "Out of patch range.";
+		}
+	}
+	catch (const char* error_message) {
+		std::cerr << "Fatal error from function GetPatch() : " << error_message << std::endl;
+		std::exit(1);
+	}
 }
 
 /***
@@ -199,26 +199,26 @@ void GOF::GetPatch(pcl::PointCloud<pcl::PointXYZRGB>& __patch, Eigen::Matrix4f& 
  * @return {*}
  */
 void GOF::GetPatches(std::vector<pcl::PointCloud<pcl::PointXYZRGB>>& __patches, std::vector<Eigen::Matrix4f>& __matrices) const {
-    try {
-        if (this->frame_patches_.size() == this->motion_vectors_.size()) {
-            __patches.resize(this->frame_patches_.size());
-            __matrices.resize(this->motion_vectors_.size());
-            for (size_t i = 0; i < this->frame_patches_.size(); i++) {
-                __patches[i].resize(this->frame_patches_[i].size());
-                for (size_t j = 0; j < this->frame_patches_[i].size(); j++) {
-                    __patches[i][j] = this->frame_patches_[i][j];
-                }
-                __matrices[i] = this->motion_vectors_[i];
-            }
-        }
-        else {
-            throw "Unmatching size between frames and matrices.";
-        }
-    }
-    catch (const char* error_message) {
-        std::cerr << "GOF is broken : " << error_message << std::endl;
-        std::exit(1);
-    }
+	try {
+		if (this->frame_patches_.size() == this->motion_vectors_.size()) {
+			__patches.resize(this->frame_patches_.size());
+			__matrices.resize(this->motion_vectors_.size());
+			for (size_t i = 0; i < this->frame_patches_.size(); i++) {
+				__patches[i].resize(this->frame_patches_[i].size());
+				for (size_t j = 0; j < this->frame_patches_[i].size(); j++) {
+					__patches[i][j] = this->frame_patches_[i][j];
+				}
+				__matrices[i] = this->motion_vectors_[i];
+			}
+		}
+		else {
+			throw "Unmatching size between frames and matrices.";
+		}
+	}
+	catch (const char* error_message) {
+		std::cerr << "GOF is broken : " << error_message << std::endl;
+		std::exit(1);
+	}
 }
 
 /***
@@ -227,7 +227,7 @@ void GOF::GetPatches(std::vector<pcl::PointCloud<pcl::PointXYZRGB>>& __patches, 
  * @return {float}
  */
 float GOF::GetResolution() const {
-    return std::max(std::max(this->max_x_ - this->min_x_, this->max_y_ - this->min_y_), this->max_z_ - this->min_z_);
+	return std::max(std::max(this->max_x_ - this->min_x_, this->max_y_ - this->min_y_), this->max_z_ - this->min_z_);
 }
 
 /***
@@ -236,7 +236,7 @@ float GOF::GetResolution() const {
  * @return {float}
  */
 float GOF::GetResolutionBinary() const {
-    return std::pow(2.0f, std::ceil(std::log2(this->GetResolution())));
+	return std::pow(2.0f, std::ceil(std::log2(this->GetResolution())));
 }
 
 /***
@@ -245,7 +245,7 @@ float GOF::GetResolutionBinary() const {
  * @return {PointXYZ}
  */
 pcl::PointXYZ GOF::GetCenter() const {
-    return pcl::PointXYZ((this->min_x_ + this->max_x_) / 2, (this->min_y_ + this->max_y_) / 2, (this->min_z_ + this->max_z_) / 2);
+	return pcl::PointXYZ((this->min_x_ + this->max_x_) / 2, (this->min_y_ + this->max_y_) / 2, (this->min_z_ + this->max_z_) / 2);
 }
 
 /***
@@ -254,8 +254,8 @@ pcl::PointXYZ GOF::GetCenter() const {
  * @return {*}
  */
 void GOF::GetMatrices(std::vector<Eigen::Matrix4f>& __matrices) const {
-    __matrices.resize(this->motion_vectors_.size());
-    for (size_t i = 0; i < this->motion_vectors_.size(); i++) {
-        __matrices[i] = motion_vectors_[i];
-    }
+	__matrices.resize(this->motion_vectors_.size());
+	for (size_t i = 0; i < this->motion_vectors_.size(); i++) {
+		__matrices[i] = motion_vectors_[i];
+	}
 }
